Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,8 @@
 #include <stdio.h>
 int main () {
-    float a , b , c ;
+    int a , b , c ;
     float d = (a+b+c)/3;
     scanf("%f %f %f",&a,&b,&c);
     printf("Average: %.2f",d);
-    
     return 0;
 }
